adicionando msg de entrada na sala / correção de erros na rota delete

+++ BackEnd/app.js
@@ -5,6 +5,7 @@ dotenv.config();
 import { MongoClient, ObjectId } from "mongodb";
 import joi from "joi";
 import dayjs from "dayjs";
+
 const app = express();
 app.use(cors());
 app.use(express.json());
@@ -17,14 +18,15 @@ mongoClient.connect().then(() => {
 });
 
 const participantsSchema = joi.object({
-  name: joi.string().required(),
+  name: joi.string().required().trim(true),
 });
 
 const messagesSchema = joi.object({
   to: joi.string().required(),
   text: joi.string().required(),
-  type: joi.string().valid("message").valid("private_message").required(),
+  type: joi.string().valid("message", "private_message").required(),
 });
+
 app.post("/participants", async (req, res) => {
   const { name } = req.body;
   const validation = participantsSchema.validate(req.body);
@@ -48,6 +50,13 @@ app.post("/participants", async (req, res) => {
       name: name,
       lastStatus: Date.now(),
     });
+    await db.collection("messages").insertOne({
+      from: name,
+      to: "Todos",
+      text: "Entrou na sala...",
+      type: "status",
+      time: dayjs().format("HH:mm:ss"),
+    });
     res.sendStatus(201);
   } catch (error) {
     res.sendStatus(500);
@@ -66,10 +75,11 @@ app.get("/participants", async (req, res) => {
 app.post("/messages", async (req, res) => {
   const { to, text, type } = req.body;
   const validation = messagesSchema.validate(req.body);
+  const { user } = req.headers;
 
   const participantsValid = await db
     .collection("participants")
-    .findOne({ name: req.headers.user });
+    .findOne({ name: user });
 
   const activeParticipants = await db
     .collection("participants")
@@ -81,7 +91,7 @@ app.post("/messages", async (req, res) => {
   }
   try {
     await db.collection("messages").insertOne({
-      from: req.headers.user,
+      from: user,
       to: to,
       text: text,
       type: type,
@@ -103,13 +113,13 @@ function filterPrivateMessage(message, user) {
 
 app.get("/messages", async (req, res) => {
   const limit = parseInt(req.query.limit);
-  /* const { user } = req.headers.user; */
+  const { user } = req.headers;
 
   try {
     const response = await db.collection("messages").find().toArray();
     const filteredMessages = response
       .splice(-limit)
-      .filter((message) => filterPrivateMessage(message, req.headers.user));
+      .filter((message) => filterPrivateMessage(message, user));
     res.send(filteredMessages);
   } catch (error) {
     res.sendStatus(500);
@@ -117,9 +127,10 @@ app.get("/messages", async (req, res) => {
 });
 
 app.post("/status", async (req, res) => {
-  const participantsValid = await db
+  const participantsValid = await db;
+  const { user } = req.headers
     .collection("participants")
-    .findOne({ name: req.headers.user });
+    .findOne({ name: user });
 
   if (!participantsValid) {
     res.sendStatus(404);
@@ -129,10 +140,7 @@ app.post("/status", async (req, res) => {
   try {
     await db
       .collection("participants")
-      .updateOne(
-        { name: req.headers.user },
-        { $set: { lastStatus: Date.now() } }
-      );
+      .updateOne({ name: user }, { $set: { lastStatus: Date.now() } });
 
     res.sendStatus(200);
   } catch (error) {
@@ -160,22 +168,26 @@ setInterval(removeInactive, 15000);
 
 app.delete("/messages/:idMessage", async (req, res) => {
   const { idMessage } = req.params;
-  console.log(idMessage);
+  const { user } = req.headers;
   try {
     const participantMessage = await db.collection("messages").findOne({
-      from: req.headers.user,
+      from: user,
       _id: ObjectId(idMessage),
     });
+
     const messageValid = await db.collection("messages").findOne({
       _id: ObjectId(idMessage),
     });
-    console.log(messageValid);
+
     if (!messageValid) {
       res.sendStatus(404);
       return;
     }
-    if (participantMessage) {
+
+    if (participantMessage && participantMessage.type !== "status") {
       await db.collection("messages").deleteOne({ _id: ObjectId(idMessage) });
+      res.sendStatus(200);
+      return;
     } else {
       res.sendStatus(401);
       return;
